requirements.txt
bcrypt
blinker
certifi
cffi
click
Flask~=2.0.1
Flask-Bcrypt
Flask-Login
Flask-Mail
Flask-SQLAlchemy
Flask-WTF
itsdangerous
Jinja2
MarkupSafe
Pillow~=8.3.1
pycparser
six
SQLAlchemy
Werkzeug~=2.0.2
WTForms~=2.3.3
email_validator
torch~=1.9.0
torchvision~=0.10.0
gdown~=4.2.0
requests~=2.27.1